+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - Fixed markup tag parsing
 - Fixed crashes when parsing invalid enums
+- Fixed error locations on variable declarations
 
 ## [0.1.0] - 2026-08-29