+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Fixed markup tag parsing
 - Fixed crashes when parsing invalid enums
 - Fixed crashes when parsing invalid use items
+- Fixed crashes when checking some reactive expressions in markup
 - Fixed error locations on variable declarations
 - Fixed misleading error on compile