+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Fixed crashes when parsing invalid enums
 - Fixed crashes when parsing invalid use items
 - Fixed crashes when checking some reactive expressions in markup
+- Fixed module resolution
 - Fixed functions treated as statements where they should be expressions
 - Fixed missing return in generated method code
 - Fixed generated markup attribute value