+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All meaningful changes to this project will be written here.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
-## Unreleased
+## [0.1.1] - 2026-09-07
 
 ### Fixed